crimson/requirements.txt
strands-agents>=1.27.0
strands-agents-tools>=0.1.0
neo4j>=5.0.0
ddtrace>=4.4.0
datadog-api-client>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
jinja2>=3.1.0
pydantic>=2.0.0
pytest>=7.0.0
